move options and keymaps into core/
