+++ dotfiles/nvim/init.lua
@@ -1,4 +1,4 @@
-require("options")
-require("keymaps")
+require("config.core.options")
+require("config.core.keymaps")
 
 require("config.lazy")
