+++ dotfiles/nvim/lua/config/core/options.lua
